pyproject.toml
[project]
name = "ylogx-backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "argon2-cffi>=25.1.0",
    "fastapi>=0.127.0",
    "openai>=2.14.0",
    "psycopg>=3.3.2",
    "psycopg-pool>=3.3.0",
    "pyjwt>=2.10.1",
    "python-dotenv>=1.2.1",
    "requests>=2.32.5",
    "ruff>=0.14.10",
    "uvicorn>=0.40.0",
]

[dependency-groups]
dev = [
    "pre-commit>=4.5.1",
    "ruff>=0.14.10",
]

[tool.ruff]
line-length = 88
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "UP"]
ignore = ["E501"]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
